feat/STAYBNB-31 fetching in server, client container, loading file, server action, parsin Decimal to number from prisma, removing amenities from basic listing
