+++ src/lib/parsers/listing.ts
@@ -1,5 +1,5 @@
 import { ListingForm } from "@/store/useListingForm";
-import { Amenity } from "../types/amenities";
+//import { Amenity } from "../types/amenities";
 import {
   CreateListingDB,
   HostListingsWithReservations,
@@ -12,14 +12,14 @@ import {
   ListingWithReservationsDB,
 } from "../types/listing";
 import { ReservedDate } from "../types/reservation";
-import { parseAmenitiesFromDB } from "./amenities";
+//import { parseAmenitiesFromDB } from "./amenities";
 import { parseReservationsFromDB } from "./reservation";
 
 export function parseListingFromDB(listingDB: ListingDB): Listing {
-  let amenities: Amenity[] = [];
+  /* let amenities: Amenity[] = [];
   if (listingDB.listing_amenities) {
     amenities = parseAmenitiesFromDB(listingDB.listing_amenities);
-  }
+  } */
   return {
     id: listingDB.id,
     hostId: listingDB.host_id,
@@ -31,7 +31,7 @@ export function parseListingFromDB(listingDB: ListingDB): Listing {
     location: listingDB.location,
     checkInTime: listingDB.check_in_time,
     checkOutTime: listingDB.check_out_time,
-    nightPrice: listingDB.night_price,
+    nightPrice: Number(listingDB.night_price),
     promotions: listingDB.promotions?.map((promo) => ({
       minNights: promo.min_nights,
       discountPercentage: promo.discount_percentage,
@@ -43,7 +43,7 @@ export function parseListingFromDB(listingDB: ListingDB): Listing {
     images: listingDB.images,
     minCancelDays: listingDB.min_cancel_days,
     status: listingDB.status,
-    amenities,
+    //amenities,
   };
 }
