+++ src/lib/parsers/reservation.ts
@@ -26,7 +26,7 @@ export function parseReservationsFromDB(reservations: ReservationDB[]): Reservat
     guests: reservation.guests,
     totalPrice: reservation.total_price,
     totalNights: reservation.total_nights,
-    nightPrice: reservation.night_price,
+    nightPrice: Number(reservation.night_price),
     discount: reservation.discount || 0,
     discountPercentage: reservation.discount_percentage || 0,
     createdAt: new Date(reservation.created_at),
@@ -49,7 +49,7 @@ export function parseResumedReservationWithListingFromDB(reservations: ResumedRe
         location: reservation.listing.location,
         propertyType: reservation.listing.property_type,
         privacyType: reservation.listing.privacy_type,
-        nightPrice: reservation.listing.night_price,
+        nightPrice: Number(reservation.listing.night_price),
         checkInTime: reservation.listing.check_in_time,
         checkOutTime: reservation.listing.check_out_time,
       },
